felicita-backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-core==2.27.1
pydantic[email]==2.10.3
python-multipart==0.0.17
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.2.0